TMGAAW-2 Create the database schema
Issue(s):
  https://workbookers.atlassian.net/browse/TMGAAW-2

To initialize the database & start coding our logic, we need to have to
list of tables below working.

  * customers
  * delivery_men
  * deliveries

The generated models and tables are within the `Db` namespace. The
pg_crypto extension is also enabled to for active record primary keys:
`config/application.rb` was tweaked to make this work.